Move ExtensionStatusChanged predicate to predicate utils
